use long long int instead

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -5,7 +5,7 @@
  * @n: number as argument
  */
 
-void print_number(long long  int n)
+void print_number(long int n)
 {
 	int num = n, tenth_power = 1;
